Refactor Facet & Add DateRangeFacet Type
